% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{createReport}
\alias{createReport}
\title{Create report of RNAseq DEG anlaysis}
\usage{
createReport(g, counts, tags, pvalues, path, pop = 400, name = "DEGreport")
}
\arguments{
\item{g}{Character vector with the group the samples belong to.}

\item{counts}{Matrix with counts for each samples and each gene.
Should be same length than pvalues vector.}

\item{tags}{Genes of DEG analysis}

\item{pvalues}{pvalues of DEG analysis}

\item{path}{path to save the figure}

\item{pop}{random genes for background}

\item{name}{name of the html file}
}
\value{
A HTML file with all figures and tables
}
\description{
This function get the count matrix, pvalues, and FC of a
DEG analysis and create a report to help to detect possible problems
with the data.
}
