% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{deg}
\alias{deg}
\alias{deg,DEGSet-method}
\title{Method to get all table stored for an specific comparison}
\usage{
deg(object, value = NULL, tidy = NULL, top = NULL, ...)

\S4method{deg}{DEGSet}(object, value = NULL, tidy = NULL, top = NULL, ...)
}
\arguments{
\item{object}{\link{DEGSet}}

\item{value}{Character to specify which table to use.}

\item{tidy}{Return data.frame, tibble or original class.}

\item{top}{Limit number of rows to return. Default: All.}

\item{...}{Other parameters to pass for other methods.}
}
\description{
Method to get all table stored for an specific comparison
}
\references{
\itemize{
\item Testing if \code{top} is whole number or not comes from:
https://stackoverflow.com/a/3477158
}
}
\author{
Lorena Pantano
}
