\name{degComb}
\alias{degComb}
\title{Get random combinations of two groups}
\usage{
degComb(g1,g2,pop)
}
\arguments{
  \item{g1}{list of samples in group 1}

  \item{g2}{list of samples in group 2}

  \item{pop}{number of combinations to be return}
}
\value{
matrix with different combinatios of two vector
}
\description{
Get random combinations of two groups
}

