% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{degComps}
\alias{degComps}
\title{Automatize the use of `results()` for multiple comparisons}
\usage{
degComps(dds, combs = NULL, contrast = NULL, alpha = 0.05,
  pairs = FALSE)
}
\arguments{
\item{dds}{[DESeq2::DESeqDataSet] obcject.}

\item{combs}{Optional vector indicating the coefficients or columns
fom `colData(dds)` to create group comparisons.}

\item{contrast}{Optional vector to specify contrast. See [DESeq2::results()].}

\item{alpha}{Numeric value used in independent filtering in [DESeq2::results()].}

\item{pairs}{Boolean to indicate whether create all comparisons or only
use the coefficient already created from [DESeq2::resultsNames()].}
}
\value{
[DEGSet] with unSrunken and Srunken results.
}
\description{
This function will extract the output of [DEseq2::results()]
and [DESeq2::lcfSrink()] for multiple comparison using:
}
\details{
* coefficients
* contrast
* Multiple columns in `colData` that match coefficients
* Multiple columns in `colData` to create all possible
contrasts
}
\examples{
library(DESeq2)
dds <- makeExampleDESeqDataSet(betaSD=1)
colData(dds)[["treatment"]] <- sample(colData(dds)[["condition"]], 12)
design(dds) <-  ~ condition + treatment
dds <- DESeq(dds)
res <- degComps(dds, combs = c("condition", 2),
              contrast = list("treatment_B_vs_A", c("condition", "A", "B")))
}
\author{
Lorena Pantano
}
