% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{degCovariates}
\alias{degCovariates}
\title{Find correlation between pcs and covariates}
\usage{
degCovariates(counts, metadata, fdr = 0.1, scale = FALSE, minPC = 5,
  correlation = "kendall", addCovDen = TRUE, legacy = FALSE,
  smart = TRUE, plot = TRUE)
}
\arguments{
\item{counts}{normalized counts matrix}

\item{metadata}{data.frame with samples metadata.}

\item{fdr}{numeric value to use as cutoff to determine
the minimum fdr to consider significant correlations
between pcs and covariates.}

\item{scale}{boolean to determine wether counts matrix should be
scaled for pca. default FALSE.}

\item{minPC}{numeric value that will be used as cutoff
to select only pcs that explain more variability than this.}

\item{correlation}{character determining the method for the
correlation between pcs and covariates.}

\item{addCovDen}{boolean. Whether to add the covariates
dendograme to the plot to see covariates relationship.
It will show \code{\link[=degCorCov]{degCorCov()}} dendograme on top of the columns of
the heatmap.}

\item{legacy}{boolean. Whether to plot the legacy version.}

\item{smart}{boolean. Whether to avoid normalization of the
numeric covariates when calculating effect size. This is not
used if \code{legacy = TRUE}. See @details for more information.}

\item{plot}{Whether to plot or not the correlation matrix.}
}
\value{
: list:
a) significantCovars, covariates with FDR below the cutoff.
b) plot, heatmap of the correlation found. * means pvalue < 0.05.
Only variables with FDR value lower than the cutoff are colored.
c) corMatrix, correlation, p-value, FDR values
for each covariate and PCA pais
d) effectsSignificantcovars: that is PCs % * absolute
correlation between covariate and PCs,
e) pcsMatrix: PCs loading for each sample
}
\description{
This function will calculate the pcs using prcomp function,
and correlate categorical and numerical variables from
metadata. The size of the dots indicates the importance of the
metadata, for instance, when the range of the values is pretty
small (from 0.001 to 0.002 in ribosimal content),
the correlation results is not important. See details to know
how this is calculated.
}
\details{
This method is adapeted from Daily et al 2017 article.
Principal components from PCA analysis are correlated with
covariates metadata. Factors are transformed to numeric variables.
Correlation is measured by cor.test function with Kendall method
by default.

The size of the dot, or effect size, indicates the importance of
the covariate based on the range of the values. Covariates
where the range is very small (like a % of mapped reads that
varies between 0.001 to 0.002) will have a very small dot size (0.1 size).
The highest value to have is 5 size.
To get to this value, each covariate is normalized using this
equation: \code{1 - min(v/max(v))},
and the minimum and maximum values are set to
0.01 and 1 respectively. 0.5 would mean there is at least
50% of difference between the minimum value and the maximum value.
Categorical variables are plot using the maximum size always, since
it is not possible to estimate the variability. By default, it
won't do \code{v/max(v)} if the values are already between 0-1 or
0-100, trying to avoid rates and percentage values.
If you want to ignore the effect size, use \code{legacy = TRUE}.
}
\examples{
data(humanGender)
library(DESeq2)
idx <- c(1:10, 75:85)
dse <- DESeqDataSetFromMatrix(assays(humanGender)[[1]][1:1000, idx],
  colData(humanGender)[idx,], design=~group)
res <- degCovariates(log2(counts(dse)+0.5), colData(dse))
res <- degCovariates(log2(counts(dse)+0.5),
  colData(dse), legacy = TRUE)
res$plot
res$scatterPlot[[1]]
}
\references{
Daily, K. et al.  Molecular, phenotypic, and sample-associated data to describe pluripotent stem cell lines and derivatives. Sci Data 4, 170030 (2017).
}
\author{
: Lorena Pantano, Kenneth Daily and Thanneer Malai Perumal
}
