% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{degFC}
\alias{degFC}
\title{get the FC for each gene between two groups}
\usage{
degFC(g1, g2, counts, popsize)
}
\arguments{
\item{g1}{list of samples in group 1}

\item{g2}{list of samples in group 2}

\item{counts}{count matrix of deregulated genes}

\item{popsize}{number of combinations to generate}
}
\value{
FC for different combinations of samples in each group for each gene
}
\description{
get the FC for each gene between two groups
}

