% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{degNcomb}
\alias{degNcomb}
\title{Get number of potential combinations of two vectors}
\usage{
degNcomb(g1, g2)
}
\arguments{
\item{g1}{list of samples in group 1}

\item{g2}{list of samples in group 2}
}
\value{
maximum number of combinations of two vectors
}
\description{
Get number of potential combinations of two vectors
}

