% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{degPlotCluster}
\alias{degPlotCluster}
\title{Plot clusters from degPattern function output}
\usage{
degPlotCluster(table, time, color = NULL, points = TRUE, boxes = TRUE,
  smooth = TRUE, lines = TRUE, facet = TRUE)
}
\arguments{
\item{table}{\code{normalized} element from \code{\link[=degPatterns]{degPatterns()}} output.}

\item{time}{column name to use in the x-axis.}

\item{color}{column name to use to color and divide the samples.}

\item{points}{Add points to the plot.}

\item{boxes}{Add boxplot to the plot.}

\item{smooth}{Add regression line to the plot.}

\item{lines}{Add gene lines to the plot.}

\item{facet}{Split figures based on cluster ID.}
}
\value{
\link{ggplot2} object.
}
\description{
This function helps to format the cluster plots from \code{\link[=degPatterns]{degPatterns()}}.
It allows to control the layers and it returns a ggplot object that
can accept more ggplot functions to allow customization.
}
\examples{
data(humanGender)
library(SummarizedExperiment)
library(ggplot2)
ma <- assays(humanGender)[[1]][1:100,]
des <- colData(humanGender)
des[["other"]] <- sample(c("a", "b"), 85, replace = TRUE)
res <- degPatterns(ma, des, time="group", col = "other", plot = FALSE)
degPlotCluster(res$normalized, "group", "other")
degPlotCluster(res$normalized, "group", "other", lines = FALSE)
}
