\name{degPR}
\alias{degPR}
\title{plot the correlation between the rank according estimator and the rank according FC}
\usage{
degPR(rank)
}
\arguments{
  \item{rank}{output from \code{degRank} function}

  \item{colors}{colour used for each gene}
}
\value{
ggplot2 object
}
\description{
plot the correlation between the rank according estimator
and the rank according FC
}

