% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{degQC}
\alias{degQC}
\title{Plot main figures showing p-values distribution and mean-variance correlation}
\usage{
degQC(pvalue, counts, groups)
}
\arguments{
\item{pvalue}{pvalues of DEG analysis}

\item{counts}{matrix with counts for each samples and each gene.}

\item{groups}{character vector with group name for each sample in the
same order than counts column names.}
}
\value{
ggplot2 object
}
\description{
This function joins the output of \code{\link[DEGreport]{degMean}}, 
\code{\link[DEGreport]{degVar}} and \code{\link[DEGreport]{degMV}} in a
single plot. See these functions for further information.
}
\examples{
library(DESeq2)
data(humanSexDEedgeR)
idx <- c(1:10, 75:85)
dse <- DESeqDataSetFromMatrix(humanSexDEedgeR$counts[1:1000, idx], 
humanSexDEedgeR$samples[idx,], design=~group)
dse <- DESeq(dse)
res <- results(dse)
degQC(res$pvalue, counts(dse, normalized=TRUE),colData(dse)$group)
}

