% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGreportSet.R
\docType{data}
\name{DEGreportSet}
\alias{DEGreportSet}
\title{List of process geuvadis data to test the package}
\format{List}
\source{
gEUvadis
}
\usage{
DEGreportSet
}
\description{
It contains gene counts matrix, group1 of samples,
group2 of samples, differential expression analysis table, 
set of genes, output from degRank function.
}
\author{
Lorena Pantano, 2014-05-31
}

