% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{degTable}
\alias{degTable}
\alias{degTable,DEGSet-method}
\title{Method to get all table stored for an specific comparison}
\usage{
degTable(object, value = NULL)

\S4method{degTable}{DEGSet}(object, value = NULL)
}
\arguments{
\item{object}{[DEGSet]}

\item{value}{Character to specify which table to use.}
}
\description{
Method to get all table stored for an specific comparison
}
\author{
Lorena Pantano
}
