% Generated by roxygen2 (4.0.0): do not edit by hand
\name{figurepvaluebyexp}
\alias{figurepvaluebyexp}
\title{Wrap figure from \code{degMean} into a Nozzle object}
\usage{
figurepvaluebyexp(pvalues, counts, out)
}
\arguments{
  \item{pvalues}{pvalues of DEG analysis}

  \item{counts}{matrix with counts for each samples and
  each gene. Should be same length than pvalues vector.}

  \item{out}{path to save the figure}
}
\value{
Nozzle object
}
\description{
Wrap figure from \code{degMean} into a Nozzle object
}

