% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{geneInfo}
\alias{geneInfo}
\title{data.frame with chromose information for each gene}
\format{data.frame}
\source{
biomart
}
\usage{
colors
}
\description{
data.frame with chromose information for each gene
}
\author{
Lorena Pantano, 2014-08-14
}

