% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cor.R
\name{geom_cor}
\alias{geom_cor}
\title{Add correlation and p-value to a \link{ggplot2} plot}
\usage{
geom_cor(mapping = NULL, data = NULL, geom = GeomCor,
  method = "spearman", position = "identity", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{method}{Method to calculate the correlation. Values are
passed to \code{\link[=cor.test]{cor.test()}}. (Spearman, Pearson, Kendall).}
}
\description{
\code{geom_cor} will add the correlatin, method and p-value to the plot
automatically guessing the position if nothing else specidfied.
family font, size and colour can be used to change the format.
}
\details{
It was integrated after reading this tutorial to extend
ggplot2 \href{https://cran.r-project.org/web/packages/ggplot2/vignettes/extending-ggplot2.html}{layers}
}
\examples{
data(humanGender)
library(ggplot2)
ggplot(as.data.frame(assay(humanGender)[1:1000,]),
       aes(x = NA20502, y = NA20504)) +
  geom_point() +
  geom_cor(method = "kendall") 
}
\seealso{
\code{\link[ggplot2:layer]{ggplot2::layer()}}
}
