% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanGender-data.R
\docType{data}
\name{humanGender}
\alias{humanGender}
\title{DGEList object for DE genes betwen Male and Females}
\format{DGEList}
\source{
gEUvadis
}
\usage{
data(humanGender)
}
\description{
DGEList object for DE genes betwen Male and Females
}
\author{
Lorena Pantano, 2017-08-37
}
