% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/humanSexDEedgeR-data.R
\docType{data}
\name{humanSexDEedgeR}
\alias{humanSexDEedgeR}
\title{edgeR object for DE genes betwen Male and Females}
\format{edgeR object}
\source{
gEUvadis
}
\usage{
humanSexDEedgeR
}
\description{
edgeR object for DE genes betwen Male and Females
}
\author{
Lorena Pantano, 2014-05-31
}

