% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot_top}
\alias{plot_top}
\title{Plot top genes allowing more variables to color and shape points}
\usage{
plot_top(dds, res, n = 9, xs = "time", group = "condition",
  batch = NULL)
}
\arguments{
\item{dds}{\link[DESeq2]{DESeqDataSet} object}

\item{res}{\link[DESeq2]{DESeqResults} object}

\item{n}{integer number of genes to plot.}

\item{xs}{character, colname in colData that will be used as X-axes}

\item{group}{character, colname in colData to color points and add different
lines for each level}

\item{batch}{character, colname in colData to shape points, normally used by 
batch effect visualization}
}
\value{
ggplot showing the expresison of the genes
}
\description{
Plot top genes allowing more variables to color and shape points
}

