% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{plotMA}
\alias{plotMA}
\alias{plotMA,DEGSet-method}
\title{MA-plot from base means and log fold changes}
\usage{
\S4method{plotMA}{DEGSet}(object, title = NULL, label_points = NULL,
  label_column = "symbol", limit = NULL, diff = 5, ...)
}
\arguments{
\item{object}{\link{DEGSet} class.}

\item{title}{\emph{Optional}. Plot title.}

\item{label_points}{Optionally label these particular points.}

\item{label_column}{Match label_points to this column in the results.}

\item{limit}{Absolute maximum to plot on the log2FoldChange.}

\item{diff}{Minimum difference between logFoldChange before and
after shrinking.}

\item{...}{Optional parameters to pass.}
}
\value{
MA-plot \link{ggplot}.
}
\description{
MA-plot addaptation to show the shrinking effect.
}
\examples{
library(DESeq2)
dds <- makeExampleDESeqDataSet(betaSD=1)
dds <- DESeq(dds)
res <- degComps(dds, contrast = list("condition_B_vs_A"))
plotMA(res[["condition_B_vs_A"]])
}
\author{
Victor Barrera

Rory Kirchner

Lorena Pantano
}
