% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{significants}
\alias{significants}
\alias{significants,DEGSet-method}
\title{Method to get the significant genes}
\usage{
significants(object, ...)

\S4method{significants}{DEGSet}(object, padj = 0.05, fc = 0, ...)
}
\arguments{
\item{object}{[DEGSet]}

\item{...}{Passed to [degTable]. Default: value = NULL.}

\item{padj}{Cutoff for the FDR column.}

\item{fc}{Cutoff for the log2FC column.}
}
\description{
Function to get the features that are significant
according to some thresholds from a [DEGSet] class.
}
\author{
Lorena Pantano
}
