\name{degBI}
\alias{degBI}
\title{Get the estimates of the FC mean from a FC distribution using bayesian inference}
\usage{
degBI(fc)
}
\arguments{
  \item{fc}{list of FC}
}
\value{
matrix with values from \link{do_estimate}
}
\description{
Get the estimates of the FC mean from a FC distribution
using bayesian inference
}

