% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{degBI}
\alias{degBI}
\title{Get the estimates of the fold change (FC) mean from a FC distribution using
bayesian inference}
\usage{
degBI(fc, iter = 1000, ncores = NULL)
}
\arguments{
\item{fc}{list of FC}

\item{iter}{number of iteration in the mcmc model}

\item{ncores}{number of cores to use}
}
\value{
matrix with values from \link{degBIcmd}
}
\description{
Get the estimates of the fold change (FC) mean from a FC distribution using
bayesian inference
}

