% Generated by roxygen2 (4.0.0): do not edit by hand
\name{degBI}
\alias{degBI}
\title{Get the estimates of the fold change (FC) mean from a FC distribution using
bayesian inference}
\usage{
degBI(fc)
}
\arguments{
  \item{fc}{list of FC}
}
\value{
matrix with values from \link{degBIcmd}
}
\description{
Get the estimates of the fold change (FC) mean from a FC
distribution using bayesian inference
}

