% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{degCovariates}
\alias{degCovariates}
\title{Find correlation between PCs and covariates}
\usage{
degCovariates(counts, metadata, fdr = 0.1, scale = FALSE, min_pc_pct = 5,
  correlation = "pearson")
}
\arguments{
\item{counts}{normalized counts matrix}

\item{metadata}{data.frame with samples metadata.}

\item{fdr}{numeric value to use as cutoff to determine
the minimum FDR to consider significant correlations
between PCs and covariates.}

\item{scale}{boolean to determine wether counts matrix should be
scaled for PCA. Default FALSE.}

\item{min_pc_pct}{numeric value that will be used as cutoff
to select only PCs that explain more variability than this.}

\item{correlation}{character determining the method for the
correlation between PCs and covariates.}
}
\value{
: list: 
a)significantCovars, 
b)plot,
c)cor_matrix, 
d)Effects.significantCovars: that is PCs pct * absolute correlation between covariate and PCs,
e)PCs_matrix: PCs loading for each sample
}
\description{
This function will calculate the PCs using prcomp function,
and correlate categorical and numerical variables from
metadata.
}
\examples{
data(humanSexDEedgeR)
library(DESeq2)
idx <- c(1:5, 75:80)
dse <- DESeqDataSetFromMatrix(humanSexDEedgeR$counts[1:1000, idx],
humanSexDEedgeR$samples[idx,], design=~group)
res <- degCovariates(log2(counts(dse)+0.5),
colData(dse), "example")
res$plot
}
\author{
: Kenneth Daily and Thanneer Malai Perumal
}
