% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{degDefault}
\alias{degDefault}
\alias{degDefault,DEGSet-method}
\title{Method to get the default table to use.}
\usage{
degDefault(object)

\S4method{degDefault}{DEGSet}(object)
}
\arguments{
\item{object}{\link{DEGSet}}
}
\description{
Method to get the default table to use.
}
\author{
Lorena Pantano
}
