% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{degMV}
\alias{degMV}
\title{Correlation of the standard desviation and the mean of the abundance of a
set of genes.}
\usage{
degMV(g1, g2, pvalues, counts)
}
\arguments{
\item{g1}{list of samples in group 1}

\item{g2}{list of samples in group 2}

\item{pvalues}{pvalues of DEG analysis}

\item{counts}{matrix with counts for each samples and each gene.
row number should be the same length than pvalues vector.}
}
\value{
ggplot2 object
}
\description{
Correlation of the standard desviation and the mean of the abundance of a
set of genes.
}
\examples{
data(DEGreportSet)
degMV(DEGreportSet$g1,DEGreportSet$g2,DEGreportSet$deg[,4],
    DEGreportSet$counts)
}

