% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{degMV}
\alias{degMV}
\title{Correlation of the standard desviation and the mean of the abundance of a
set of genes.}
\usage{
degMV(group, pvalues, counts, sign = 0.01)
}
\arguments{
<<<<<<< HEAD
\item{group}{character vector with group name for each sample in the
same order than counts column names.}

\item{pvalues}{pvalues of DEG analysis}

\item{counts}{matrix with counts for each samples and each gene.}

\item{sign}{defining the cutoff to label significant features.
=======
\item{group}{Character vector with group name for each sample in the
same order than counts column names.}

\item{pvalues}{pvalues of DEG analysis.}

\item{counts}{Matrix with counts for each samples and each gene.}

\item{sign}{Defining the cutoff to label significant features.
>>>>>>> master
row number should be the same length than pvalues vector.}
}
\value{
ggplot2 object
}
\description{
Correlation of the standard desviation and the mean of the abundance of a
set of genes.
}
\examples{
data(DEGreportSet)
degMV(c(rep("M", length(DEGreportSet$g1)), rep("F", length(DEGreportSet$g2))),
<<<<<<< HEAD
      DEGreportSet$deg[,4],
=======
      DEGreportSet$deg[, 4],
>>>>>>> master
      DEGreportSet$counts)
}
