% Generated by roxygen2 (4.0.0): do not edit by hand
\name{degObj}
\alias{degObj}
\title{Create a deg object that can be used to plot expression values
    at shiny server:runGist(9930881)}
\usage{
degObj(counts,design,outfile)
}
\arguments{
  \item{counts}{output from get_rank function}

  \item{design}{colour used for each gene}

  \item{outfile}{file that will contain the object}
}
\value{
R object to be load into vizExp
}
\description{
Create a deg object that can be used to plot expression
values at shiny server:runGist(9930881)
}

