% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{degObj}
\alias{degObj}
\title{Create a deg object that can be used to plot expression values
    at shiny server:runGist(9930881)}
\usage{
degObj(counts, design, outfile)
}
\arguments{
<<<<<<< HEAD
\item{counts}{output from get_rank function}

\item{design}{colour used for each gene}

\item{outfile}{file that will contain the object}
}
\value{
R object to be load into vizExp
=======
\item{counts}{Output from get_rank function.}

\item{design}{Colour used for each gene.}

\item{outfile}{File that will contain the object.}
}
\value{
R object to be load into vizExp.
>>>>>>> master
}
\description{
Create a deg object that can be used to plot expression values
    at shiny server:runGist(9930881)
}
\examples{
data(DEGreportSet)
<<<<<<< HEAD
de = data.frame(row.names=colnames(DEGreportSet$counts),
=======
de = data.frame(row.names=colnames(DEGreportSet$counts), 
>>>>>>> master
sex = c(rep("M", length(DEGreportSet$g1)),
        rep("F", length(DEGreportSet$g2))))
degObj(DEGreportSet$counts, de, NULL)
}
