% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{degObj}
\alias{degObj}
\title{Create a deg object that can be used to plot expression values
at shiny server:runGist(9930881)}
\usage{
degObj(counts, design, outfile)
}
\arguments{
\item{counts}{Output from get_rank function.}

\item{design}{Colour used for each gene.}

\item{outfile}{File that will contain the object.}
}
\value{
R object to be load into vizExp.
}
\description{
Create a deg object that can be used to plot expression values
at shiny server:runGist(9930881)
}
\examples{
data(humanGender)
library(SummarizedExperiment)
degObj(assays(humanGender)[[1]], colData(humanGender), NULL)
}
