% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genePlots.R
\name{degPlot}
\alias{degPlot}
\title{Plot top genes allowing more variables to color and shape points}
\usage{
degPlot(dds, xs, res = NULL, n = 9, genes = NULL, group = NULL,
  batch = NULL, metadata = NULL, ann = c("external_gene_name", "symbol"),
  slot = 1L, xsLab = xs, groupLab = group, batchLab = batch)
}
\arguments{
\item{dds}{\link[DESeq2]{DESeqDataSet} object or SummarizedExperiment
or Matrix or data.frame.}

\item{xs}{Character, colname in colData that will be used as X-axes.}

\item{res}{\link[DESeq2]{DESeqResults} object.}

\item{n}{Integer number of genes to plot.}

\item{genes}{Character of gene names matching rownames of count data.}

\item{group}{Character, colname in colData to color points and add different
lines for each level.}

\item{batch}{Character, colname in colData to shape points, normally used by
batch effect visualization.}

\item{metadata}{Metadata in case dds is a matrix.}

\item{ann}{Columns in rowData (if available) used to print gene names.}

\item{slot}{Name of the slot to use to get count data.}

\item{xsLab}{Character, alternative label for x-axis (default: same as xs).}

\item{groupLab}{Character, alternative label for group (default: same as group).}

\item{batchLab}{Character, alternative label for batch (default: same as batch).}
}
\value{
ggplot showing the expresison of the genes
}
\description{
Plot top genes allowing more variables to color and shape points
}
