% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{degPR}
\alias{degPR}
\title{plot the correlation between the rank according estimator and
    the rank according FC}
\usage{
degPR(rank,colors)
}
\arguments{
  \item{rank}{output from \code{degRank} function}

  \item{colors}{colour used for each gene}
}
\value{
ggplot2 object
}
\description{
plot the correlation between the rank according estimator
and the rank according FC
}
\examples{
data(DEGreportSet)
degPR(DEGreportSet$rank)
}

