% Generated by roxygen2 (4.0.1): do not edit by hand
\name{degRank}
\alias{degRank}
\title{Get rank data frame with best score on the top}
\usage{
degRank(g1, g2, counts, fc, popsize, iter = 1000, ncores = NULL)
}
\arguments{
  \item{g1}{list of samples in group 1}

  \item{g2}{list of samples in group 2}

  \item{counts}{count matrix for each gene and each sample
  that is deregulated}

  \item{fc}{list of FC of deregulated genes. Should be same
  length than counts \code{row.names}}

  \item{popsize}{number of combinations to generate}

  \item{iter}{number of iteration in the mcmc model}

  \item{ncores}{number of cores to use}
}
\value{
data frame with the output of \link{degBIcmd} for each gene
}
\description{
Get rank data frame with best score on the top
}
\examples{
data(DEGreportSet)
degRank(DEGreportSet$g1,DEGreportSet$g2,
    DEGreportSet$counts[DEGreportSet$detag[1:5],],
    DEGreportSet$deg[DEGreportSet$detag[1:5],1],400,500)
}

