% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGreport-package.R
\docType{package}
\name{DEGreport-package}
\alias{DEGreport}
\alias{DEGreport-package}
\title{Deprecated functions in package DEGreport}
\description{
These functions are provided for compatibility with older versions
of DEGreport only and will be defunct at the next release.
}
\details{
The following functions are deprecated and will be made defunct;
use the replacement indicated below:
\itemize{
\item degRank, degPR, degBIcmd, degBI, degFC, degComb, degNcomb: DESeq2::lcfShrink.
This function was trying to avoid big FoldChange
in variable genes. There are other methods nowadays like lcfShrink function.
DEGreport
}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://lpantano.github.io/DEGreport/}
  \item Report bugs at \url{https://github.com/lpantano/DEGreport/issues}
}

}
\author{
\strong{Maintainer}: Lorena Pantano \email{lorena.pantano@gmail.com}

Other contributors:
\itemize{
  \item John Hutchinson [contributor]
  \item Victor Barrera [contributor]
  \item Mary Piper [contributor]
  \item Radhika Khetani [contributor]
  \item Kenneth Daily [contributor]
  \item Thanneer Malai Perumal [contributor]
  \item Rory Kirchner [contributor]
  \item Michael Steinbaugh [contributor]
}

}
