% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DEGreportSet.R
\docType{data}
\name{DEGreportSet}
\alias{DEGreportSet}
\title{list object for DE genes betwen Male and Females}
\format{matrix,list,list and matrix}
\source{
gEUvadis
}
\usage{
DEGreportSet
}
\description{
list of objects containgin counts matrix,g1,g2 and edgeR
glmfit object
}
\author{
Lorena Pantano, 2014-05-31
}

