% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DEGSet}
\alias{DEGSet}
\alias{DEGSet-class}
\title{DEGSet}
\description{
S4 class to store data from differentially expression analysis.
It should be compatible with different package and stores the information
in a way the methods will work with all of them.
}
\details{
For now supporting only \code{\link[DESeq2:results]{DESeq2::results()}} output.
Use constructor \code{\link[=degComps]{degComps()}} to create the object.

The list will contain one element for each comparison done.
Each element has the following structure:
\itemize{
\item DEG table
\item Optional table with shrunk Fold Change when it has been done.
}

To access the raw table use \code{deg(dgs, "raw")}, to access the
shrunken table use \code{deg(dgs, "shrunk")} or just \code{deg(dgs)}.
}
\examples{
library(DESeq2)
dds <- makeExampleDESeqDataSet(betaSD = 1)
colData(dds)[["treatment"]] <- sample(colData(dds)[["condition"]], 12)
design(dds) <-  ~ condition + treatment
dds <- DESeq(dds)
res <- degComps(dds, combs = c("condition"))
deg(res[[1]])
deg(res[[1]], tidy = "tibble")
}
\author{
Lorena Pantano
}
