% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{degVar}
\alias{degVar}
\title{Distribution of pvalues by standard desviation range}
\usage{
degVar(pvalues, counts)
}
\arguments{
\item{pvalues}{pvalues of DEG analysis}

\item{counts}{matrix with counts for each samples and each gene.
row number should be the same length than pvalues vector.}
}
\value{
ggplot2 object
}
\description{
Distribution of pvalues by standard desviation range
}
\examples{
data(DEGreportSet)
degVar(DEGreportSet$deg[,4],DEGreportSet$counts)
}

