% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{degVB}
\alias{degVB}
\title{Distribution of the standard desviation of
DE genes compared to the background}
\usage{
degVB(tags, g1, g2, counts, pop = 400)
}
\arguments{
\item{tags}{List of genes that are DE.}

\item{g1}{List of samples in group 1.}

\item{g2}{List of samples in group 2.}

\item{counts}{matrix with counts for each samples and each gene.
Should be same length than pvalues vector.}

\item{pop}{Number of random samples taken for background comparison.}
}
\value{
ggplot2 object
}
\description{
Distribution of the standard desviation of
DE genes compared to the background
}
\examples{
data(humanGender)
library(DESeq2)
idx <- c(1:10, 75:85)
dds <- DESeqDataSetFromMatrix(assays(humanGender)[[1]][1:1000, idx],
  colData(humanGender)[idx,], design=~group)
dds <- DESeq(dds)
res <- results(dds)
degVB(row.names(res)[1:20], colData(dds)[["group"]][1:10],
  colData(dds)[["group"]][11:20], counts(dds, normalized = TRUE))
}
