% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report.R
\name{figurerank}
\alias{figurerank}
\title{Wrap figure from \code{plotrank} into a Nozzle object}
\usage{
figurerank(tab, out, colors)
}
\arguments{
  \item{tab}{table from \link{degRank}}

  \item{out}{path to save the figure}

  \item{colors}{colors for each gene}
}
\value{
Nozzle object
}
\description{
Wrap figure from \code{plotrank} into a Nozzle object
}

