% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanSexDEedgeR-data.R
\docType{data}
\name{humanSexDEedgeR}
\alias{humanSexDEedgeR}
\title{DGEList object for DE genes betwen Male and Females}
\format{DGEList}
\source{
gEUvadis
}
\usage{
humanSexDEedgeR
}
\description{
DGEList object for DE genes betwen Male and Females
}
\author{
Lorena Pantano, 2014-05-31
}

