% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{significants}
\alias{significants}
\alias{significants,DEGSet-method}
\alias{significants,DESeqResults-method}
\alias{significants,TopTags-method}
\title{Method to get the significant genes}
\usage{
significants(object, ...)

\S4method{significants}{DEGSet}(object, padj = 0.05, fc = 0,
  direction = NULL, ...)

\S4method{significants}{DESeqResults}(object, padj = 0.05, fc = 0,
  direction = NULL, ...)

\S4method{significants}{TopTags}(object, padj = 0.05, fc = 0,
  direction = NULL, ...)
}
\arguments{
\item{object}{\link{DEGSet}}

\item{...}{Passed to \link{deg}. Default: value = NULL.
Value can be 'raw', 'shrunken'.}

\item{padj}{Cutoff for the FDR column.}

\item{fc}{Cutoff for the log2FC column.}

\item{direction}{Whether to take down/up/ignore. Valid arguments are
down, up and NULL.}
}
\description{
Function to get the features that are significant
according to some thresholds from a \link{DEGSet},
\link[DESeq2:DESeqResults]{DESeq2::DESeqResults} and \link[edgeR:topTags]{edgeR::topTags}.
}
\author{
Lorena Pantano
}
