% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tablerank}
\alias{tablerank}
\title{Create table for Nozzle report}
\usage{
tablerank(tab, out)
}
\arguments{
  \item{tab}{table from \code{\link{degRank}}}

  \item{out}{path to save the figure}
}
\value{
Nozzle object
}
\description{
Create table for Nozzle report
}

